Rename to Asset Cache Cleaner and bump version to 1.2.0

+++ Editor/AssetCacheCleaner/AssetCacheCleanerWindow.cs
@@ -7,20 +7,24 @@ using System.Linq;
 using UnityEditor;
 using UnityEngine;
 
-namespace AssetCacheRemover {
-    internal class AssetCacheRemoverWindow : EditorWindow {
-        const string _editorPrefsAssetStoreCachePathKey = "AssetCacheRemover_AssetStoreCachePath";
+namespace AssetCacheCleaner {
+    internal class AssetCacheCleanerWindow : EditorWindow {
+        const string _editorPrefsAssetStoreCachePathKey = "AssetCacheCleaner_AssetStoreCachePath";
+        const string _editorPreShowConfirmationPathKey = "AssetCacheCleaner_ShowConfirmation";
 
         string _cachePath = string.Empty;
+        bool _showConfirmation = true;
+
         readonly SortedList<string, (string Publisher, string Path)> _assets = new();
         Vector2 _scrollPosition = Vector2.zero;
 
-        [MenuItem("Tools/Asset Cache Remover")]
+        [MenuItem("Tools/Asset Cache Cleaner")]
         static void ShowWindow() {
-            var window = GetWindow<AssetCacheRemoverWindow>();
-            window.minSize = new Vector2(500, 600);
-            window.titleContent = new GUIContent("Asset Cache Remover");
+            var window = GetWindow<AssetCacheCleanerWindow>();
+            window.minSize = new Vector2(600, 200);
+            window.titleContent = new GUIContent("Asset Cache Cleaner");
             window.UpdateAssetStoreCachePath();
+            window.UpdateShowConfirmation();
             window.UpdatePackagesList();
         }
 
@@ -42,6 +46,9 @@ namespace AssetCacheRemover {
 #endif
         }
 
+        void UpdateShowConfirmation() =>
+            _showConfirmation = EditorPrefs.GetBool(_editorPreShowConfirmationPathKey, true);
+
         void UpdatePackagesList() {
             if (!Directory.Exists(_cachePath))
                 return;
@@ -53,10 +60,11 @@ namespace AssetCacheRemover {
                 foreach (var subDirectory in subDirectories) {
                     var files = Directory.GetFiles(subDirectory);
                     foreach (var file in files) {
-                        if (file.EndsWith(".unitypackage")) {
-                            var asset = Path.GetFileNameWithoutExtension(file);
-                            _assets[asset] = (publisher, file);
-                        }
+                        if (!file.EndsWith(".unitypackage"))
+                            continue;
+
+                        var asset = Path.GetFileNameWithoutExtension(file);
+                        _assets[asset] = (publisher, file);
                     }
                 }
             }
@@ -74,20 +82,37 @@ namespace AssetCacheRemover {
                 UpdatePackagesList();
             }
 
+            var newShowConfirmation = GUILayout.Toggle(_showConfirmation, "Show confirmation dialog");
+            if (newShowConfirmation != _showConfirmation) {
+                _showConfirmation = newShowConfirmation;
+                EditorPrefs.SetBool(_editorPreShowConfirmationPathKey, _showConfirmation);
+            }
+
             EditorGUILayout.EndVertical();
 
             EditorGUILayout.Space(15);
             _scrollPosition = EditorGUILayout.BeginScrollView(_scrollPosition);
 
             foreach (var (key, value) in _assets.ToList()) {
+                var asset = key + " by " + value.Publisher;
                 EditorGUILayout.BeginHorizontal();
+                GUILayout.Label(asset);
 
                 if (GUILayout.Button("Remove")) {
-                    File.Delete(value.Path);
-                    _assets.Remove(key);
+                    var shouldRemove = !_showConfirmation || EditorUtility.DisplayDialog(
+                        "Confirm Asset Removal",
+                        $"Are you sure you want to remove\n\n   {asset}\n\n" +
+                        "from the Asset Store cache folder on your disk?\n\n" +
+                        "You will need to re-download the asset from the Asset Store whenever you need it again.\n\n" +
+                        "Assets in your project(s) will remain safe and untouched.",
+                        "Remove", "Cancel");
+
+                    if (shouldRemove) {
+                        File.Delete(value.Path);
+                        _assets.Remove(key);
+                    }
                 }
 
-                GUILayout.Label(key + " by " + value.Publisher);
                 EditorGUILayout.EndHorizontal();
             }
